Lab2/requirements.txt
python == 3.12.4
numpy == 1.26.4
Pillow == 10.3.0
matplotlib == 3.8.4